refactor(python/timestamp): flatten package structure and keep tests in tests/ directory
