+++ samples/python/extensions/timestamp/__init__.py
@@ -1,4 +1,4 @@
-from timestamp_ext.core import TIMESTAMP_FIELD, URI, TimestampExtension
+from core import TIMESTAMP_FIELD, URI, TimestampExtension
 
 
 __all__ = [
