+++ samples/python/extensions/timestamp/core.py
