+++ samples/python/extensions/timestamp/py.typed
@@ -1,0 +1,1 @@
+# Marker file for PEP 561.
